utils/jwt/token_test.go
// Copyright © 2020 Bin Liu <bin.liu@enmotech.com>

package jwt

import (
	"testing"
)

func TestGenerateToken(t *testing.T) {
	type args struct {
		userID   int64
		username string
		expire   int64
	}
	tests := []struct {
		name    string
		args    args
		want    string
		wantErr bool
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			got, err := GenerateToken(tt.args.userID, tt.args.username, tt.args.expire)
			if (err != nil) != tt.wantErr {
				t.Errorf("GenerateToken() error = %v, wantErr %v", err, tt.wantErr)
				return
			}
			if got != tt.want {
				t.Errorf("GenerateToken() got = %v, want %v", got, tt.want)
			}
		})
	}
}
